+++ src/mchammer/__init__.py
@@ -1,21 +1,20 @@
 """MCHammer package."""
 
-from mchammer.atom import Atom
-from mchammer.bond import Bond
-from mchammer.collapser import Collapser
-from mchammer.mc_operations import (
-    get_bond_vector,
+from mchammer._internal.atom import Atom
+from mchammer._internal.bond import Bond
+from mchammer._internal.collapser import Collapser
+from mchammer._internal.mc_operations import (
     rotate_molecule_by_angle,
     rotation_matrix_arbitrary_axis,
     test_move,
     translate_atoms_along_vector,
     translate_molecule_along_vector,
 )
-from mchammer.molecule import Molecule
-from mchammer.optimizer import Optimizer
-from mchammer.radii import get_radius
-from mchammer.results import MCStepResult, Result, StepResult
-from mchammer.utilities import get_atom_distance
+from mchammer._internal.molecule import Molecule
+from mchammer._internal.optimizer import Optimizer
+from mchammer._internal.radii import get_radius
+from mchammer._internal.results import MCStepResult, Result, StepResult
+from mchammer._internal.utilities import get_atom_distance, get_bond_vector
 
 __all__ = [
     "Atom",
