+++ tests/stk_reproduction/utilities.py
@@ -13,8 +13,6 @@ def get_mch_bonds(state: stk.ConstructedMolecule):  # noqa: ANN201
         The state of the molecule under construction.
 
     Yields:
-    ------
-    :class:`MCHammer.Bond`
         A bond in the molecule.
 
     """
@@ -36,8 +34,6 @@ def get_long_bond_ids(state: stk.ConstructedMolecule):  # noqa: ANN201
         The state of the molecule under construction.
 
     Yields:
-    ------
-    :class:`tuple` of :class:`int`
         A pair of atom ids that identify a bond to be optimized.
 
     """
